requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-dotenv==1.0.1 
snowflake-connector-python==3.6.0
requests==2.31.0 